refactor : Input import 위치 수정
